Moved bootstrap and kernel to lib folder
